Make distinction between created and received invoice history items
Before this commit, the invoice history view would only show "received" as a
history entry. This commit changes that so that a "created" history item is
shown when the author of an invoice is looking at it in the history view, and
"received" is still shown when the recipient is looking at the history.

Closes #1764