Fix missing Business Type field
The "business type" field was not showing up for businesses that were of type
"Sole Proprietorship". This was because the id's for business types are
numerical and started at 0, which was the id for "Sole Proprietorship". Because
of JavaScript's falsiness/truthiness rules, and id of 0 was a falsy value whereas
all other numbers would be truthy.

This commit changes the check from one for truthiness to one that checks that the
type is numeric. It also adds a fallback message if the type is numeric but isn't
recognized by the system.