Remove duplicate invoice put from cron
When the ScheduleInvoiceCron cron job runs to pay all scheduled
invoices, it modifies the relevant invoice objects to mark them as paid.
However, it also puts the relevant transaction to the transactionDAO,
which also puts the relevant invoice to the invoiceDAO. Since invoiceDAO
is wrapped with a HistoryDAO decorator, two history records were being
added to the invoiceHistoryDAO when the cron job ran instead of one.
This caused an error in the invoice history view for invoices paid via
the cron job, showing an extra history item. This commit removes the put
to the invoice object in the ScheduleInvoiceCron cron job and leaves the
modification of the invoice to UpdateInvoiceTransactionDAO.

Closes #2172