Fix invitation security bug
Since we were setting the sender ID on the frontend, a tech-savvy user
would have been able to set the id manually to another account and thus
spoof sending invites from other accounts. This commit fixes that
possibility by setting the `createdBy` property on the backend instead
based on the user that created the invitation.