Use PublicUserInfoDAO on invoiceDAO
Make use of the reusable PublicUserInfoDAO on the invoiceDAO to populate
the payer and payee fields with public user information about those
users. This removes the need for the properties `payerName` and
`payeeName` on the Invoice model, so this commit removes those
properties and replaces all references to them with references to
paye{r,e}.label() instead.