Fix GUI bug when verifying bank accounts
Since the verifyAccount method was on the inner class
BankAccountTableView, when it pushed a new view onto the stack, the new
view was keeping some of the UI elements from the parent's UI (the
parent being BankAccountsView. By moving the method to the parent class
we avoid this problem because `this` in the method will now refer to the
parent class instead of the inner class and the stack will completely
replace the parent view instead of keeping part of it and only replacing
the part containing the inner class.