Make contactDAO branch off of bareUserDAO
When we create an invoice, we want to be able to set the payee to either
a User or a Contact.

Therefore, a User must never have the same id as a Contact.

To solve this problem, we store Users and Contacts in the same DAO, so
they both hit the same SequenceNumberDAO and therefore will never have
id collisions.

One note is that we don't want contact to simply extend userDAO, because
we don't necessarily want contacts to have the same decorators as users.
Therefore we created a "pointer" called bareUserDAO that points to a
very deep point in the userDAO chain of decorators.

This commit makes contactDAO branch off of bareUserDAO instead of being
separate, solving the id problem and letting us set the payee id on an
invoice without ambiguity.