Remove duplicate properties from the contact modal view
This way we don't need to write code all over the place that copies back
and forth between the view and data. Just put the properties into the
view and let FOAM handle the two-way data binding like it was built to
do. This also lets us reuse the validation written on the model instead
of copying it in a million different views. It also lets us blur the
line between editing and creating, because either way we have a form
with fields that are bound to the data, whether the data is empty or
not.