Delete ReadDisabledUserDAO
It was created specifically to not show deleted or frozen (disabled)
businesses in the select business view in Ablii, but requirements
changed after this was implemented. The new requirements state that
frozen businesses should show up in the list but in a disabled state
instead of not showing up in the list at all. Since that can be
accomplished using an mlang, this decorator can be removed from the
chain of decorators on localUserDAO and since it's not reusable, it can
be deleted.