Change entity FObjectProperty to entityId and entityDaoKey
Since the entity FObjectProperty stores a copy of the entity data on complianceHistory object and has no knowledge of its DAO service, updating the original entity is not possible. Replace it with entityId and entityDaoKey properties to reference the entity id and DAO key instead so that DAO service can be retrieved and used for data update.