Fixes a security concern, see description
Fixes a security issue where, if a user signs up for Ablii, invites a
signing officer, and goes to the user management screen without logging
out, then when the new signing officer signs up, the original user's
group in their session wasn't getting updated until they logged out and
back in. So they could remain an admin in the business as long as they
stayed signed in. This commit fixes that by updating their group in
their session as soon as the signing officer signs up.