Update list when contact added
Needed to change to a proper DAO instead of copying values into an
ArrayDAO. This way it listens for changes on the underlying DAO and
updates when changes are made. Now when you add a contact it will update
the list and show the newly added contact in it without needing to
refresh.