Fix bug where 2FA screen had left margin
The problem was that the custom CSS from the theme was being applied
before the 2FA screen instead of after, which is how it was working
before.

I refactored the code to not require the custom CSS anyway. The views
should be handling that, not CSS from the group or theme.