Fix bug that occurs when submitting the business profile
If a user selects 1 or 2 people that own 25% or more of the business but
does not check "I am one of these owners", then an error is thrown when
they submit the business onboarding.

Was happening because when we clone the BusinessOnboarding object before
submitting it, the order in which the properties were cloned would cause
the `owner1` property to be cleared after it had been set during the
cloning process.

The fix in this commit is to only clear that property when the box has
been checked.