Remove lookups in type to improve performance and fix search
Since the 'type' property was transient, it wasn't available on the
server side and thus couldn't be used in mlangs, meaning search wouldn't
work for that property.

When the type was dow jones or identity mind, each row in the table
would have to do an additional lookup to populate the table cell for the
type property. Actually storing the value removes this requirement as
well, which should have a significant positive impact on performance.

When I tested, loading the initial few pages of the compliance items
table took a whopping 335 requests, 320 of which were finds to populate
the type cells.

Another issue I noticed is that the count is being requested thrice, and
each select (of which there are three) is happening four times. I'll
have to look into why we're duplicating requests so many times. Just
wanted to note that here though since the 320 number above should really
be a fourth of that.