Remove decorator that was throwing error (#8772)
This is a cheap workaround to get user deletion working again. The
decorator removed in this commit was throwing a NullPointerException
because ApprovableAwareDAO was further down the decorator chain and was
returning null from `put_`.

Rather than trying to fix the decorator to play nice with
LifecycleAware and ApprovableAware, I realized that this decorator is
onnly relevant to the "act as another user" feature that's not currently
used in Liquid anyway, so we can just remove it for now as a
time-sensitive fix. We're only removing it in the liquid deployment
journals, not in the base journals.