Remove clearing owner1 from init and signingOfficer postSet
so that owner1 info will not be cleared out when backoffice users view/edit the onboarding.

The issue was that the `init` and `signingOfficer` postSet in BusinessOnboarding are async functions and FOAM already loaded the data when those functions are called thus clearing owner1 essentially mean resetting owner1 info for no reason and it also triggers the event to reset the "I certify..." checkbox because the owner1 info was changed.