Refactor to not call BlobService.put directly.
Rather, rely on the File model/DAO, so that the blob/file id is setup correctly and
the File model can determine where to save the file data.
Also, correct the FileService 'file' resolution logic as it was testing for
File.dataString or File.data, both of which are empty when the file has been
saved to the file system as in independent file (not journalled).

NP-2843 #time 6h