Execute applyTo once per session.
Each client call, on each request, whether DIG, HTTP, Socket,
performs session context setup via applyTo. applyTo is the most
expensive logic of context setup, but the results of applyTo do
not change between calls, so it makes sense to reuse the
context setup.  Auth checks are independent of this.
Also, if the system were to reconfigure a theme, for example,
the existing sessions could be invalidated or the the 'context'
of each be reset to force applyTo to run on next client call.

This change in a medusa cluster provides a 20% performance gain
for 'ping' test on a 1Gbs network.