Throw MedusaException during Replaying
On mediator restart, existing admin logins will attempt http operations,
resulting in SessionServerBox 'put' operations on the sessionDAO,
which in turn cause 'put' from the secondaries to the primary.
But during replay, either the primary is not found, or the request
is sent, and the calling thread blocks in ReplayingDAO.  The browser
sends multiple requests and can potentially block a large number of
threads, causing resource issues for replaying.