Refactory acquisition of AppConfig
Replace group.getAppConfig(x) with Session.applyTo(x).get("appConfig").
Session.applyTo performs a number of tests to ensure a valid and useful AppConfig.
Also, inject email template parameters link and appLink in the Email and Notification
classes, rather than having all callers of notification having to aquire and set.
Future work could split the path components of a url, so the caller only provides the
path and is not concerned with the domain. i.e. {{link}}/images/...