MedusaEntry.objectId - handle multipart keys
objectid was originally storageTransient and thus not persisted on nodes.
With compaction objectId was retained - now persisted.
ObjectId is type Object and multiplart ids are models.
On parsing the multipart id class must exists, but this is not the
case with nodes, as they are not built with the entire application source.
UnknownFObjectParser is used to parse the objectId property,
allowing all existing logic to remain, by simply treating the json
model as a string.