CronSchedule - change getNextScheduledTime to use curren time to day, rather than start of day, when lastRunDate is null. .
getNextScheduledTime calculates the next time to run a cron job.
The 'next' time follows the 'lastRunDate'.
Cron.lastRunDate is storageTransient to avoid journal entries for
every cron job run. After system start, all 'lastRunDate' values are
null, and the first call to getNextScheduledTime has to select a
suitable 'from' date to perform calculations.  Hithertoo this value
was now.startofday (0:00).  In a Medusa cluster when the Primary
changes, the new Primary's table of cron jobs still has null for
all lastRunDate properties (as this value is also not clustered -
see StorageOptional). As a result the new Primary getNextScheduledTime
calls will calculate from startofday, which is incorrect and
results in cron jobs running twice in one day, for example.
Changing the default 'from' time to the current time of day
resolves this issue. Only the hour test case required adjustment
as it assumed the initial 'from' time was start of day. All other
24 test cases passed without adjustment. Note that it was the
behaviour for an instant to run all jobs for the day regardless
of when it was started.  It is not the responsibility of the
operater to verify and manually run any jobs which should have
run earlier than when the server started.