Fix missing arg1 on Not.create
It was passing binary predicate to the create method so the arg1 was taken from the binary predicate.
Not.arg1 is a predicate but Binary.arg1 is an expression, this casuses the predicate property adapt to try to create a new predicate and failed because Predicate is an interface.

Changing 'type' of PredicateProperty to AbstractPredicate causes java compilation because we code to Predicate interface instead of the abstract class.
We could add 'javaType' to point back to the interface but the deviation will become a technical debt for the future.