Register FileSystemStorage and Storage in the context
FileSystemStorage is used for store and forward, blobstore services,
etc... which require disk access.

Storage is the default storage system for JDAO and others that might
need access to resource files inside the jar file.

For jar deployment, the Storage register will be a ResourceStorage.
For development build, the Storage register will be same as
FileSystemStorage.